split logger responsibilities in different modules
