+++ packages/logger/src/timer.ts
@@ -1,0 +1,12 @@
+export type Timer = { start(): { stop(): number } };
+
+export const defaultTimer = {
+  start() {
+    const start = Date.now();
+    return {
+      stop() {
+        return Date.now() - start;
+      }
+    };
+  }
+};
